+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2023-05-17
+
+- Added TTS Simplex Streaming
+- Added TTS Duplex Streaming
+
 ## [1.0.1] - 2023-03-28
 
 We implemented basic SDK features.